Move model tests to the right place.
