+++ hyperspy/tests/model/test_model.py
@@ -1,9 +1,118 @@
+import os
+
 import numpy as np
 import numpy.testing
 import nose.tools
 
 from hyperspy.hspy import *
+from hyperspy.components import Gaussian
 
+
+class TestModel:
+
+    def setUp(self):
+        s = signals.Spectrum(np.empty(1))
+        m = create_model(s)
+        self.model = m
+
+    def test_access_component_by_name(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        nose.tools.assert_is(m["test"], g2)
+
+    def test_access_component_by_index(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        nose.tools.assert_is(m[1], g2)
+
+    def test_component_name_when_append(self):
+        m = self.model
+        gs = [Gaussian(), Gaussian(), Gaussian()]
+        m.extend(gs)
+        nose.tools.assert_is(m['Gaussian'], gs[0])
+        nose.tools.assert_is(m['Gaussian_0'], gs[1])
+        nose.tools.assert_is(m['Gaussian_1'], gs[2])
+
+    @nose.tools.raises(ValueError)
+    def test_several_component_with_same_name(self):
+        m = self.model
+        gs = [Gaussian(), Gaussian(), Gaussian()]
+        m.extend(gs)
+        m[0]._name = "Gaussian"
+        m[1]._name = "Gaussian"
+        m[2]._name = "Gaussian"
+        m['Gaussian']
+
+    @nose.tools.raises(ValueError)
+    def test_no_component_with_that_name(self):
+        m = self.model
+        m['Voigt']
+
+    @nose.tools.raises(ValueError)
+    def test_component_already_in_model(self):
+        m = self.model
+        g1 = Gaussian()
+        m.extend((g1, g1))
+
+    def test_remove_component(self):
+        m = self.model
+        g1 = Gaussian()
+        m.append(g1)
+        m.remove(g1)
+        nose.tools.assert_equal(len(m), 0)
+
+    def test_remove_component_by_index(self):
+        m = self.model
+        g1 = Gaussian()
+        m.append(g1)
+        m.remove(0)
+        nose.tools.assert_equal(len(m), 0)
+
+    def test_remove_component_by_name(self):
+        m = self.model
+        g1 = Gaussian()
+        m.append(g1)
+        m.remove(g1.name)
+        nose.tools.assert_equal(len(m), 0)
+
+    def test_get_component_by_name(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        nose.tools.assert_is(m._get_component("test"), g2)
+
+    def test_get_component_by_index(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        nose.tools.assert_is(m._get_component(1), g2)
+
+    def test_get_component_by_component(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        nose.tools.assert_is(m._get_component(g2), g2)
+
+    @nose.tools.raises(ValueError)
+    def test_get_component_wrong(self):
+        m = self.model
+        g1 = Gaussian()
+        g2 = Gaussian()
+        g2.name = "test"
+        m.extend((g1, g2))
+        m._get_component(1.2)
 
 class TestModelFitBinned:
 
@@ -81,7 +190,6 @@ class TestModelFitBinned:
     @nose.tools.raises(ValueError)
     def test_wrong_method(self):
         self.m.fit(method="dummy")
-
 
 class TestModelWeighted:
 
@@ -174,7 +282,6 @@ class TestModelWeighted:
         self.m.fit(fitter="leastsq", method="ls")
         nose.tools.assert_almost_equal(self.m.red_chisq.data, 3.37700055)
 
-
 class TestModelScalarVariance:
 
     def setUp(self):
@@ -225,7 +332,6 @@ class TestModelScalarVariance:
         self.m.fit(fitter="leastsq", method="ls")
         nose.tools.assert_almost_equals(self.m.red_chisq.data, 0.86206965)
 
-
 class TestModelSignalVariance:
 
     def setUp(self):
@@ -251,7 +357,6 @@ class TestModelSignalVariance:
                                         0.79693355673230915)
         nose.tools.assert_almost_equals(self.m.red_chisq.data[1],
                                         0.91453032901427167)
-
 
 class TestMultifit:
